Add ldd and bdd modules to merc_symbolic to split it up further
