+++ crates/aterm/src/storage/gc_mutex.rs
@@ -20,9 +20,9 @@ pub struct GcMutex<T> {
     inner: UnsafeCell<T>,
 }
 
+// SAFETY: Sharing GcMutex<T> across threads gives shared &T access, so T must be Send.
 unsafe impl<T: Send> Send for GcMutex<T> {}
 // SAFETY: Sharing &GcMutex<T> across threads gives shared &T access, so T must be Sync.
-unsafe impl<T: Send + Sync> Sync for GcMutex<T> {}
 unsafe impl<T: Send + Sync> Sync for GcMutex<T> {}
 
 impl<T> GcMutex<T> {
