+++ crates/symbolic/src/bdd/format.rs
